Add Lab6

+++ Lab05/books/src/main/java/pl/edu/pwr/ztw/books/controller/AuthorsController.java
@@ -6,6 +6,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import pl.edu.pwr.ztw.books.model.Author;
 import pl.edu.pwr.ztw.books.service.IAuthorsService;
+import pl.edu.pwr.ztw.books.service.IBooksService;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -17,10 +18,19 @@ public class AuthorsController {
     @Autowired
     IAuthorsService authorsService;
 
+    @Autowired
+    IBooksService booksService;
+
     @RequestMapping(value = "/authors", method = RequestMethod.GET)
-    @Operation(summary = "Get all authors")
-    public ResponseEntity<Object> getAuthors() {
-        return new ResponseEntity<>(authorsService.getAuthors(), HttpStatus.OK);
+    @Operation(summary = "Get all authors with pagination")
+    public ResponseEntity<Object> getAuthors(
+            @RequestParam(defaultValue = "0") int page,
+            @RequestParam(defaultValue = "5") int size
+    ) {
+        return ResponseEntity.ok()
+                .header("X-Total-Count", String.valueOf(authorsService.getAuthorsCount()))
+                .header("Access-Control-Expose-Headers", "X-Total-Count")
+                .body(authorsService.getAuthors(page, size));
     }
 
     @RequestMapping(value = "/authors/{id}", method = RequestMethod.GET)
@@ -56,6 +66,9 @@ public class AuthorsController {
     @RequestMapping(value = "/authors/{id}", method = RequestMethod.DELETE)
     @Operation(summary = "Delete author")
     public ResponseEntity<Object> deleteAuthor(@Parameter(description = "Author ID", example = "1") @PathVariable("id") int id) {
+        if (!booksService.getBooksByAuthor(id).isEmpty()) {
+            return new ResponseEntity<>("Nie można usunąć autora z przypisanymi książkami", HttpStatus.CONFLICT);
+        }
         boolean deleted = authorsService.deleteAuthor(id);
         if (deleted) {
             return new ResponseEntity<>("Author deleted successfully", HttpStatus.OK);
